scale_test/requirements.txt
Babel==2.1.1
cliff==1.15.0
cmd2==0.6.8
ecdsa==0.13
httplib2==0.9.2
iso8601==0.1.10
lxml==3.3.3
ncclient
netaddr==0.7.18
oslo.config==2.5.0
paramiko==1.15.3
pbr==0.11.0
prettytable==0.7.2
pyparsing==2.0.3
python-keystoneclient==0.9.0
python-neutronclient==2.3.4
python-novaclient==2.17.0
pytz==2015.6
PyYAML==3.11
requests==2.8.1
simplejson==3.8.0
six==1.10.0
stevedore==1.9.0
unicodecsv==0.14.1
